+++ core/trainers/tdm_trainer.py
@@ -12,65 +12,53 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Training use fluid with one node only.
+Training use fluid with DistributeTranspiler
 """
-
-from __future__ import print_function
+import os
+import time
 import logging
-
 import numpy as np
 import paddle.fluid as fluid
-from paddlerec.core.trainers.single_trainer import SingleTrainer
+from paddle.fluid.incubate.fleet.parameter_server.distribute_transpiler import fleet
 from paddlerec.core.utils import envs
+from paddlerec.core.trainers.general_trainer import GeneralTrainer, EngineMode
 
 logging.basicConfig(format="%(asctime)s - %(levelname)s - %(message)s")
 logger = logging.getLogger("fluid")
 logger.setLevel(logging.INFO)
-special_param = [
-    "TDM_Tree_Travel", "TDM_Tree_Layer", "TDM_Tree_Info", "TDM_Tree_Emb"
-]
+special_param = ["TDM_Tree_Travel", "TDM_Tree_Layer", "TDM_Tree_Info"]
 
 
-class TDMSingleTrainer(SingleTrainer):
+class TDMTrainer(GeneralTrainer):
+
     def startup(self, context):
+        logger.info("Run TDM Trainer Startup Pass")
+        if self.engine == EngineMode.SINGLE:
+            self._single_startup(context)
+        else:
+            self._cluster_startup(context)
+
+        context['status'] = 'train_pass'
+
+    def _single_startup(self, context):
         namespace = "train.startup"
-        load_persistables = envs.get_global_env("single.load_persistables",
-                                                False, namespace)
+        load_tree_from_numpy = envs.get_global_env("single.load_tree_from_numpy",
+                                                   False, namespace)
+        self._exe.run(self._startup_program)
+        if load_tree_from_numpy:
+            logger.info("load tree from numpy")
 
-        persistables_model_path = envs.get_global_env(
-            "single.persistables_model_path", "", namespace)
+            self.tree_layer_path = envs.get_global_env("tree.tree_layer_path", "",
+                                                       namespace)
 
-        load_tree = envs.get_global_env("tree.load_tree", False, namespace)
+            self.tree_travel_path = envs.get_global_env("tree.tree_travel_path",
+                                                        "", namespace)
 
-        self.tree_layer_path = envs.get_global_env("tree.tree_layer_path", "",
-                                                   namespace)
+            self.tree_info_path = envs.get_global_env("tree.tree_info_path", "",
+                                                      namespace)
 
-        self.tree_travel_path = envs.get_global_env("tree.tree_travel_path",
-                                                    "", namespace)
-
-        self.tree_info_path = envs.get_global_env("tree.tree_info_path", "",
-                                                  namespace)
-
-        self.tree_emb_path = envs.get_global_env("tree.tree_emb_path", "",
-                                                 namespace)
-
-        save_init_model = envs.get_global_env("single.save_init_model", False,
-                                              namespace)
-        init_model_path = envs.get_global_env("single.init_model_path", "",
-                                              namespace)
-        self._exe.run(fluid.default_startup_program())
-
-        if load_persistables:
-            # 从paddle二进制模型加载参数
-            fluid.io.load_persistables(
-                executor=self._exe,
-                dirname=persistables_model_path,
-                main_program=fluid.default_main_program())
-            logger.info("Load persistables from \"{}\"".format(
-                persistables_model_path))
-
-        if load_tree:
-            # covert tree to tensor, set it into Fluid's variable.
+            self.tree_emb_path = envs.get_global_env("tree.tree_emb_path", "",
+                                                     namespace)
             for param_name in special_param:
                 param_t = fluid.global_scope().find_var(param_name).get_tensor(
                 )
@@ -80,13 +68,46 @@ class TDMSingleTrainer(SingleTrainer):
                 else:
                     param_t.set(param_array.astype('int32'), self._place)
 
+        load_paddle_model = envs.get_global_env("single.load_paddle_model",
+                                                False, namespace)
+        assert load_tree_from_numpy != load_paddle_model, "Please Don't use load_tree_from_numpy & load_paddle_model at the same time"
+
+        if load_paddle_model:
+            # 从paddle二进制模型加载参数
+            warmup_model_path = envs.get_global_env(
+                "warmup_model_path", None, "train.trainer")
+            assert warmup_model_path != None, "set train.trainer.warmup_model_path for loading model"
+            fluid.io.load_persistables(
+                executor=self._exe,
+                dirname=warmup_model_path,
+                main_program=self._main_program)
+            logger.info("Load persistables from \"{}\"".format(
+                warmup_model_path))
+
+        save_init_model = envs.get_global_env("single.save_init_model",
+                                              False, namespace)
+        init_model_path = envs.get_global_env(
+            "single.init_model_path", "", namespace)
         if save_init_model:
             logger.info("Begin Save Init model.")
             fluid.io.save_persistables(
                 executor=self._exe, dirname=init_model_path)
             logger.info("End Save Init model.")
 
-        context['status'] = 'train_pass'
+    def _cluster_startup(self, context):
+        warmup_model_path = envs.get_global_env(
+            "warmup_model_path", None, "train.trainer")
+        assert warmup_model_path != None, "set train.trainer.warmup_model_path for loading model"
+        self._exe.run(self._startup_program)
+
+        def is_tdm_tree_var(var):
+            res = var.name in special_param
+            return res
+
+        fluid.io.load_vars(self._exe,
+                           dirname=warmup_model_path,
+                           main_program=self._main_program,
+                           predicate=is_tdm_tree_var)
 
     def _tdm_prepare(self, param_name):
         if param_name == "TDM_Tree_Travel":
@@ -98,9 +119,6 @@ class TDMSingleTrainer(SingleTrainer):
         elif param_name == "TDM_Tree_Info":
             info_array = self._tdm_info_prepare()
             return info_array
-        elif param_name == "TDM_Tree_Emb":
-            emb_array = self._tdm_emb_prepare()
-            return emb_array
         else:
             raise " {} is not a special tdm param name".format(param_name)
 
@@ -110,13 +128,6 @@ class TDMSingleTrainer(SingleTrainer):
         logger.info("TDM Tree leaf node nums: {}".format(travel_array.shape[
             0]))
         return travel_array
-
-    def _tdm_emb_prepare(self):
-        """load tdm tree param from npy/list file"""
-        emb_array = np.load(self.tree_emb_path)
-        logger.info("TDM Tree node nums from emb: {}".format(emb_array.shape[
-            0]))
-        return emb_array
 
     def _tdm_layer_prepare(self):
         """load tdm tree param from npy/list file"""
